门店查询关联门店等级与门店类型

+++ xpaas-channel-center-channel/src/main/java/com/deepexi/channel/controller/StoreTypeController.java
@@ -45,7 +45,7 @@ public class StoreTypeController {
     }
 
     @GetMapping("/{id}")
-    public Payload<StoreTypeVO> detail(@PathVariable(value = "id", required = true) Integer  id) {
+    public Payload<StoreTypeVO> detail(@PathVariable(value = "id", required = true) Long  id) {
         StoreTypeDTO storeTypeDTO = storeTypeService.detail(id);
         if(storeTypeDTO == null){
             return new Payload();
